Fix test code for Serial. Should be in Arduino.h header file.
